+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## v1.0.9 (TBA)
 
 * Removed call to `Pow.Ecto.Context.repo/1`
+* Fixed bug with exception raised in `Pow.Ecto.Schema.normalize_user_id_field_value/1` when calling `Pow.Ecto.Context.get_by/2` with a non binary user id
+* Fixed bug with exception raised in `Pow.Ecto.Schema.normalize_user_id_field_value/1` when calling `Pow.Ecto.Context.authenticate/2` with a non binary user id
+* `Pow.Ecto.Context.authenticate/2` now returns nil if user id or password is nil
 
 ## v1.0.8 (2019-05-24)